1204_2350

+++ run_SACv3.py
@@ -55,7 +55,7 @@ def hyperparameters(env_name="Hopper-v4"):
     parser.add_argument('--eval_step', '-es', default=5000, type=int, help='Frequency in performance evaluation')
     parser.add_argument('--max_step', '-ms', default=6e6, type=int, help='Maximum training step')
     parser.add_argument('--model_train_start_step', '-mtss', default=3e6, type=int)
-    parser.add_argument('--reg_weight', '-rw', default=5.0e-12, type=int)
+    parser.add_argument('--reg_weight', '-rw', default=5.0e-12, type=float)
 
     # estimate a model dynamics
     parser.add_argument('--develop-mode', '-dm', default='imn', help="none, mn, imn")
